fix issue with gcc 4.1 compile warnings

+++ starka/src/lib/starling_common/test/align_path_test.cpp
@@ -43,7 +43,7 @@ BOOST_AUTO_TEST_CASE( test_align_path_cigar_conversion ) {
 
     path_t apath;
     cigar_to_apath("10I10M10D10M10S",apath);
-    BOOST_CHECK_EQUAL(apath.size(),5);
+    BOOST_CHECK_EQUAL(static_cast<int>(apath.size()),5);
 
     // test round-trip:
     test_single_cigar_conversion("10I10M2S20M2I10M10D10M");
@@ -56,7 +56,7 @@ BOOST_AUTO_TEST_CASE( test_align_path_ref_length ) {
 
     path_t apath;
     cigar_to_apath("2I10M10D4I10M10N10M3S",apath);
-    BOOST_CHECK_EQUAL(apath_ref_length(apath),50);
+    BOOST_CHECK_EQUAL(static_cast<int>(apath_ref_length(apath)),50);
 }
 
 
@@ -64,7 +64,7 @@ BOOST_AUTO_TEST_CASE( test_align_path_read_length ) {
 
     path_t apath;
     cigar_to_apath("2I10M10D4I10M10N10M3S",apath);
-    BOOST_CHECK_EQUAL(apath_read_length(apath),39);
+    BOOST_CHECK_EQUAL(static_cast<int>(apath_read_length(apath)),39);
 }
